Widget: break out reusable header component
